IMDB_Statistics_1
The class name is same as the jar file name.
1. Given a input zipcode, find all the user-ids that belongs to that
zipcode. You must take the input zipcode in command line.  [For example,
if the input zipcode is 75252 then you need to find all users who lives
in 75252]?
2. Find top 10 average rated movies with descending order of rating ?
3.First map-reduce job will find the average rating of each movies and
write into disk. And the second map-reduce job will take the previous
output as the input and find the top 10 from them.